+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+* Refactor `IMAGES` service extending.
+    - Removed `ImageCacheProxy` API.
+    - Added similar `ImagesExtension` API.
+    - All *proxy* behavior can be easy converted to new API.
+
 * **Breaking** Length units `max`, `min` and `abs` now take by value.
 * Fix memory leak in `Var::flat_map_vec` when the returned item var is a retained clone.